#902 move TrainingManager.calculateWeeklyTrainingForPlayer to Player class
